backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
python-dotenv==1.0.0
structlog==24.1.0
aiobotocore==2.9.0
websockets==12.0
redis==5.0.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-boto3sqs==0.43b0